% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg_methods.R
\name{coef.linreg}
\alias{coef.linreg}
\title{Coefficients Method for linreg Objects}
\usage{
\method{coef}{linreg}(object, ...)
}
\arguments{
\item{object}{An object of class "linreg".}

\item{...}{Additional arguments (not used).}
}
\value{
A named numeric vector of coefficients.
}
\description{
Returns the coefficients of a linreg object as a named vector.
}
