% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg_methods.R
\name{coef.ridgereg}
\alias{coef.ridgereg}
\title{Coefficient Method for ridgereg Class}
\usage{
\method{coef}{ridgereg}(object, ...)
}
\arguments{
\item{object}{An object of class "ridgereg".}

\item{...}{Additional arguments (not used).}
}
\value{
A numeric vector of coefficients.
}
\description{
Coefficient Method for ridgereg Class
}
