% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinReg.R
\name{linreg}
\alias{linreg}
\title{Linear Regression Model using QR Decomposition}
\usage{
linreg(formula, data)
}
\arguments{
\item{formula}{A formula object describing the model to be fitted.}

\item{data}{A data.frame containing the variables in the model.}
}
\value{
An object of class "linreg" containing the model results.
}
\description{
This function performs a multiple linear regression using QR decomposition.
It calculates the regression coefficients, fitted values, residuals, degrees of freedom,
residual variance, variance of the regression coefficients, t-values, and p-values for each coefficient.
}
\examples{
data(mtcars)
model <- linreg(mpg ~ wt + hp, data = mtcars)
summary(model)

}
