% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-doc.R
\name{LRpkg}
\alias{LRpkg}
\alias{LRpkg-package}
\title{Linear Regression Models}
\description{
A package for linear regression analysis using QR decomposition.
}
\details{
LRpkg: A Package for Linear Regression Models

This package provides functionalities to handle linear regression models
using QR decomposition. It implements an S3 class to manage special methods
such as print(), plot(), resid(), pred(), coef(), and summary(). The package
focuses on the implementation of linear regression using linear algebra techniques
and allows users to easily interpret regression results.

Key features of the package include:
\itemize{
\item Calculation of regression coefficients, fitted values, residuals, and their variances.
\item Comprehensive summary statistics including t-values and p-values for each coefficient.
\item Methods for visualizing regression results, such as residuals vs fitted values and scale-location plots.
\item User-friendly interface to extract residuals and predicted values.
\item Example usage provided in the package documentation and vignettes.
}

A vignette is included to demonstrate how to conduct a simple regression analysis
using a dataset included in the package.
}
