% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg_methods.R
\name{plot.linreg}
\alias{plot.linreg}
\title{Plot Method for linreg Objects}
\usage{
\method{plot}{linreg}(x, ...)
}
\arguments{
\item{x}{An object of class "linreg".}

\item{...}{Additional arguments (not used).}
}
\description{
Creates residuals vs fitted and scale-location plots for linreg objects.
}
