% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg_methods.R
\name{pred}
\alias{pred}
\title{Prediction Method for linreg Class}
\usage{
pred(object, ...)
}
\arguments{
\item{object}{An object of class "linreg".}

\item{...}{Additional arguments (not used).}
}
\value{
A numeric vector of fitted values.
}
\description{
Prediction Method for linreg Class
}
