% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg_methods.R
\name{pred.ridgereg}
\alias{pred.ridgereg}
\title{Prediction Method for ridgereg Class}
\usage{
\method{pred}{ridgereg}(object, ...)
}
\arguments{
\item{object}{An object of class "ridgereg".}

\item{...}{Additional arguments (not used).}
}
\value{
A numeric vector of fitted values.
}
\description{
Prediction Method for ridgereg Class
}
