% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg_methods.R
\name{print.linreg}
\alias{print.linreg}
\title{Print Method for linreg Objects}
\usage{
\method{print}{linreg}(x, ...)
}
\arguments{
\item{x}{An object of class "linreg".}

\item{...}{Additional arguments (not used).}
}
\description{
Prints the coefficients of a linreg object.
}
