% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg_methods.R
\name{print.ridgereg}
\alias{print.ridgereg}
\title{Print Method for ridgereg Class}
\usage{
\method{print}{ridgereg}(x, ...)
}
\arguments{
\item{x}{An object of class "ridgereg".}

\item{...}{Additional arguments (not used).}
}
\description{
Print Method for ridgereg Class
}
