% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg_methods.R
\name{resid.linreg}
\alias{resid.linreg}
\title{Residuals Method for linreg Objects}
\usage{
\method{resid}{linreg}(object, ...)
}
\arguments{
\item{object}{An object of class "linreg".}

\item{...}{Additional arguments (not used).}
}
\value{
A numeric vector of residuals.
}
\description{
Returns the residuals of a linreg object.
}
