% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg.R
\name{ridgereg}
\alias{ridgereg}
\title{Ridge Regression using QR Decomposition}
\usage{
ridgereg(formula, data, lambda)
}
\arguments{
\item{formula}{A formula specifying the model to be fitted. The response variable
should be on the left-hand side and the predictor variables on the right-hand side.}

\item{data}{A data frame containing the variables specified in the formula.}

\item{lambda}{A non-negative numeric value representing the ridge penalty term.}
}
\value{
A list containing the following components:
\item{coefficients}{A numeric vector of estimated coefficients for the model.}
\item{fitted_values}{A numeric vector of fitted values from the model.}
\item{residuals}{A numeric vector of residuals from the model.}
\item{df}{Degrees of freedom for the model.}
\item{residual_variance}{Estimated variance of the residuals.}
\item{beta_variance}{Estimated variance-covariance matrix of the coefficients.}
\item{t_values}{A numeric vector of t-values for the estimated coefficients.}
\item{p_values}{A numeric vector of p-values corresponding to the t-values.}
\item{lambda}{The value of the ridge penalty used in the model.}
\item{formula}{The formula used to fit the model.}
\item{data}{The data frame used in the analysis.}
\item{gcv}{Generalized Cross-Validation score for the model.}
}
\description{
This function performs ridge regression on a given dataset using QR decomposition
for solving the normal equations, while also normalizing the covariates.
}
\examples{
data(mtcars)
ridge_result <- ridgereg(mpg ~ wt + hp, data = mtcars, lambda = 1)
print(ridge_result$coefficients)

}
