% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Liu_Theme.R
\name{theme_liu}
\alias{theme_liu}
\title{Liu Theme for ggplot2}
\usage{
theme_liu()
}
\value{
A ggplot2 theme object that can be added to ggplot2 plots.
}
\description{
This function creates a custom theme for ggplot2 visualizations with a
blue background and white text elements. The theme is designed to enhance
the visual appeal and readability of plots, particularly when representing
data related to flight delays or any other data visualizations that benefit
from a blue color scheme.
}
