% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_airport_delays.R
\name{visualize_airport_delays}
\alias{visualize_airport_delays}
\title{Visualize Average Flight Delays by Airport Location}
\usage{
visualize_airport_delays()
}
\value{
A ggplot2 object representing the scatter plot of average flight delays.
}
\description{
This function creates a scatter plot visualizing the average flight delays
for different airports using the \code{nycflights13} dataset. The average delay
is calculated by taking the mean of the arrival and departure delays.
The resulting plot displays the airports' geographical locations, with
point sizes and colors indicating the average delay times.
}
\examples{
library(LRpkg)
library(nycflights13)
visualize_airport_delays()

}
