% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg_methods.R
\name{pred.ridgereg}
\alias{pred.ridgereg}
\title{pred.ridgereg: Generate predictions from a ridge regression model}
\usage{
\method{pred}{ridgereg}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{ridgereg}, which contains the ridge regression model,
the fitted coefficients, and the original data used to build the model.}

\item{newdata}{A data frame containing new observations for the predictors. If provided,
predictions will be generated for these new observations. If NULL (the default),
the function will use the original data from \code{object}.}

\item{...}{Additional arguments (currently not used).}
}
\value{
A vector of predicted values. If \code{newdata} is provided, the predictions are based on
this data. Otherwise, the predictions are based on the original data in \code{object}.
}
\description{
This function generates predictions based on a ridge regression model created with \code{ridgereg}.
It can either use the original data (if \code{newdata} is not provided) or make predictions
on a new dataset.
}
