% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg_methods.R
\name{pred.ridgereg}
\alias{pred.ridgereg}
\title{Predict Fitted Values from Ridge Regression Model}
\usage{
\method{pred}{ridgereg}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{A ridge regression model object returned by the \code{ridgereg} function.}

\item{newdata}{A data frame containing the new data for which predictions are to be made.
If \code{NULL} (default), the original data used to fit the model will be used.}

\item{...}{Additional arguments to be passed to methods.}
}
\value{
A numeric vector of fitted values corresponding to the new data.
}
\description{
This function predicts the fitted values from a ridge regression model. It can handle new data
or use the data originally used to fit the model if no new data is provided. The function
standardizes the predictor variables using the mean and standard deviation from the original data.
}
